changes done in price-table-two
